// src/interfaces/UserDataInterfaces.ts
export interface UserData {
    kind: Kind;
    data: UserDataObject;
}
// ...
export enum Kind {
    T1 = "t1",
    T3 = "t3",
}